+++ server/testingHelpers.js
@@ -42,8 +42,12 @@ function sensorsVitalFactory(overrides = {}) {
 
 async function sensorsVitalDBFactory(db, overrides = {}) {
   // prettier-ignore
+  const location = {
+    locationid: overrides.locationid || 'myLocation',
+  }
+
   const sensorVital = await db.logSensorsVital(
-    overrides.locationid !== undefined ? overrides.locationid : 'myLocation',
+    overrides.location !== undefined ? overrides.location : location,
     overrides.missedDoorMessages !== undefined ? overrides.missedDoorMessages : 0,
     overrides.isDoorBatteryLow !== undefined ? overrides.isDoorBatteryLow : false,
     overrides.doorLastSeenAt !== undefined ? overrides.doorLastSeenAt : new Date('2022-01-03T04:05:06'),