+++ CHANGELOG.md
@@ -11,6 +11,13 @@ the code was deployed.
 
 ## [Unreleased]
 
+### Changed
+
+- Changed `logSensorsVital` and `getMostRecentSessionWithDeviceid` to use `location` instead of `locationid` (CU-86dqq93dp).
+- Changed `getMostRecentSessionWithDeviceid`to `getMostRecentSessionWithDevice` to reflect the change of the function (CU-86dqq93dp).
+- Changed `sensorsVitalDBFactory` to use location instead of locationid in the logSensorsVital call (CU-86dqq93dp).
+- Modified several test cases in `handleHeartbeatTest.js` to reflect the changes of `logSensorsVital`, `getMostRecentSessionWithDevice`, and `sensorsVitalDBFactory` (CU-86dqq93dp).
+
 ## [10.10.0] - 2024-06-04
 
 ### Added